🔨 Add Twilio to the rotation of SMS backends
This commit adds our newly crafted `TwilioClient` to the rotation of
`sms_clients` registered with the `Notifications-api` application.

Furthermore, we update our `apply_dev_defaults` script to add a new
Twilio `ProviderDetails` to the database, required for tracking the
provider's status (enabled or disabled) as well as its priority within
the round-robin.

https://app.asana.com/0/1199707043388412/1199945098442987