🔥 Delete (seemingly) nonsensical test
In a recent PR we added support for a new backend SMS provider, Twilio.
As part of this new integrations we needed to update the
`update_notification_status_by_reference` method which, until now, has
only dealt with delivery callbacks for email delivery, hence it would
only process callbacks for emails in the "sending" or "pending" status.

The final delivery state for a Twilio SMS is "delivered", hence we added
support for callbacks about messages in the "sent" state, causing one of
the unit specs in our test-suite to fail.

Based on the test name, it looks like the author was trying to test
delivery notification behaviour for an (international?) message that
doesn't support delivery receipts? The implementation of the test,
however, doesn't seem to relate to the tests name.

This commit removes the problematic test from our test-suite since it's
unclear what's actually been exercised or verified based on the test's
implementation.

https://app.asana.com/0/1199707043388412/1199945098442987