🐳 Move DB creation within Docker Postgres init
Until now, creation of Postgres database has been handled by the
`scripts/bootstrap.sh` script, though this depended on the presence of
the `createbd` command on the user's local path (which will only be
present if the user has `Postgresql` setup locally).

We wish to offload all application dependencies into Docker, including
`Postgresql`, hence local dependency on the `createdb` command is an
issue.

This commit offloads creation of the `notification_api` and
`notification_test` databases to the initialization process of the
`Postgres` Docker container itself, thus removing the local dependency
on the `createdb` utility.