🔥 Remove superfluous display_results function
This commit removes a `display_result` function that isn't used as part
of the `scripts/bootstrap.sh` script. This function appears to be
intended for displaying the results of a test-suite run, though the
function is already defined in `scripts/run_tests.sh`, hence it's
needlessly duplicated in our bootstrap script.

This commit removes a `display_result` function that isn't used as part
of the `scripts/bootstrap.sh` script.