✨ Seed environment.sh if not already present
Successful running of the `notifications-api` app requires a handful of
environment variables to be set on the host. Managing these environment
variables is done through an `environment.sh` file that's updated by the
user and automatically `source`d when needed, for example, in our
`scripts/run_app.sh` script.

Until now, we've relied on the user manually creating the
`environment.sh` file, since it'll (eventually) contain sensitive data
and therefore shouldn't be tracked in version control.

This commit updates our `scripts/bootstrap.sh` script to check for the
presence of the `environment.sh` file and, when absent, creates the file
with sensible default values.