🐛 Fix DB seeding if environment variable unset
In a previous commit we added a new database migration to set the email
address of the seeded user based on the `NOTIFY_EMAIL_DOMAIN`
environment variable.

Unfortunately, in our CI environment, the `NOTIFY_EMAIL_DOMAIN` variable
in unset, hence the migration process fails.

This commit adds a default domain of "example.com" if the
`NOTIFY_EMAIL_DOMAIN` variable hasn't been set.