✨ Add template for go-live request admin alerts
Whenever a service requests to "go-live", we need to notify the Catalyst
admins so they can review (and hopefully approve) the request. The
original GOV UK implementation uses the Zendesk API to create a support
ticket, though we're unsure it Catalyst use Zendesk internally, hence
we've opted to send an email alert, instead.

This commit adds a new email template to the `templates` table, which is
then used to inform admins of newly submitted go-live requests.