✨ Send notifications for new go-live requests
Whenever a service requests to "go-live", we need to notify the Catalyst
admins so they can review (and hopefully approve) the request. The
original GOV UK implementation uses the Zendesk API to create a support
ticket, though we're unsure it Catalyst use Zendesk internally, hence
we've opted to send an email alert, instead.

This commit updates our API to detect if/when a service has requested to
go-live and, when required, uses a recently added email template to send
a notification email to a support inbox, advising Catalyst admins to
review the new request.