🐛 Update ORGANISATION_TYPES to support charities
In a recent commit we updated the `Service` signup form to support new
types of `Service`, namely "Charity" and "Community Interest Company".
Unfortunately, When making these changes, we overlooked a list of
supported organisation types in the API codebase, leading to validation
failures when attempting to create new `Organisation` records.

This commit updates the list of supported `Organisation` types to
include charities and community interest companies.