🐛 Stop sending of go-live email after activation
Whenever a user requests for their `Service` to "go-live", the
`Service`, the user's database ID is saved against the `Service` in the
`go_live_user` column. Until now, we've used the change in this database
column to trigger an admin notification email, highlighting that a
"go-live" request needs to be reviewed.

During testing, we've spotted that the `Service.go_live_user` field is
reset to `None` once a request has been reviewed and a `Service` has
been toggled to "Live". This (second) change in `Service.go_live_user`
triggers a second admin notification email to be (incorrectly) sent.

This commit updates our email sending logic to only send an admin
notification email when `go_live_user` is passed through as a request
parameter, rather than watching for changes in the database.